src/Linguistic_features_classifier/requirements.txt
pandas
numpy
plotly
scikit-learn
spacy
pyphen
gitpython
pt_core_news_sm @ https://github.com/explosion/spacy-models/releases/download/pt_core_news_sm-3.5.0/pt_core_news_sm-3.5.0-py3-none-any.whl
